Update README.md
This portfolio website was developed to showcase the personal and professional profile of Khaled Almalki, a senior IT student at Taif University specializing in application development and data analysis. The website features a modern and responsive design built using HTML, CSS, and JavaScript, and serves as a central hub to present key skills, academic background, and featured projects in machine learning, survey automation, and smart emergency prediction.